pyproject.toml
[project]
name = "krakenings-discord-bot"
version = "0.1.0"
description = "discord bot for Krakenings schedule"
readme = "README.md"
requires-python = "~=3.13.0"
dependencies = [
  "discord.py~=2.4.0",
  "aiohttp~=3.11.11",
  "python-dateutil~=2.8.2",
  "nhl-stats-client>=1.0.0",
  "click>=8.1.8",
  # only require audioop-lts on Python 3.13 and higher.
  'audioop-lts ; python_version>="3.13"',
  "environ-config>=24.1.0",
  "tomli>=2.2.1",
  "rich>=13.9.4",
  "pydantic>=2.10.6",
  "httpx>=0.28.1",
  "pydantic-extra-types>=2.10.2",
]
license = {text = "MIT"}

[tool.ruff.lint.isort]
known-first-party = ["krakenings_discord_bot"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.3.4",
    "ruff>=0.9.4",
]

# [tool.hatch.build.targets.wheel]
# packages = ["src/krakenings_discord_bot"]
